chore: minor improvements to tests (#91)
