+++ tests/integration/conftest.py
@@ -81,15 +81,6 @@ def fixture_test_db_flask_image(pytestconfig: Config):
     return test_flask_image
 
 
-@pytest.fixture(scope="module", name="test_db_flask_image")
-def fixture_test_db_flask_image(pytestconfig: Config):
-    """Return the --test-flask-image test parameter."""
-    test_flask_image = pytestconfig.getoption("--test-db-flask-image")
-    if not test_flask_image:
-        raise ValueError("the following arguments are required: --test-db-flask-image")
-    return test_flask_image
-
-
 @pytest.fixture(scope="module", name="expressjs_app_image")
 def fixture_expressjs_app_image(pytestconfig: Config):
     """Return the --expressjs-app-image test parameter."""
