+++ tests/unit/fastapi/constants.py
@@ -1,6 +1,9 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+"""FastAPI constants."""
+
+
 DEFAULT_LAYER = {
     "services": {
         "fastapi": {
