+++ tox.ini
@@ -85,7 +85,7 @@ commands =
     black --target-version=py310 --check --diff {[vars]all_path}
     mypy {[vars]src_path} {[vars]legacy_src_path}
     pydocstyle {[vars]src_path} {[vars]legacy_src_path}
-    pylint {[vars]src_path} {[vars]legacy_src_path} --min-similarity-lines=10
+    pylint {[vars]src_path} {[vars]legacy_src_path} {[vars]tst_path} --min-similarity-lines=10
 
 [testenv:static]
 description = Run static analysis tests
